requirements.txt
elasticsearch==8.11.1
python-dateutil==2.8.2
elastic-apm==6.18.0 